pyproject.toml
[tool.poetry]
name = "eros"
version = "0.1.0"
description = "An image and video tagger"
authors = ["Jules"]

[tool.poetry.dependencies]
python = "^3.9"
numpy = "^1.26.0"
Pillow = "^10.0.0"
onnxruntime = "^1.16.0"
ffmpeg-python = "^0.2.0"
rich = "^13.7.1"
textual = "^0.65.1"
pydantic = "^2.0.0"
imagehash = "^4.3.1"
click = "^8.1.7"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
